+++ src/sample/Positions.java
@@ -2,7 +2,7 @@ package sample;
 
 public class Positions {
 
-    private double pk1;
+    private int pk1;
     private String batch_uid;
     private String tittle;
     private String description;
@@ -10,18 +10,18 @@ public class Positions {
     public Positions() {
     }
 
-    public Positions(double pk1, String batch_uid, String tittle, String description) {
+    public Positions(int pk1, String batch_uid, String tittle, String description) {
         this.pk1 = pk1;
         this.batch_uid = batch_uid;
         this.tittle = tittle;
         this.description = description;
     }
 
-    public double getPk1() {
+    public int getPk1() {
         return pk1;
     }
 
-    public void setPk1(double pk1) {
+    public void setPk1(int pk1) {
         this.pk1 = pk1;
     }
